+++ CHANGELOG.md
@@ -11,6 +11,7 @@
  - Added `\Wizaplace\SDK\Catalog\AttributeVariant::getSeoDescription`
  - Added `\Wizaplace\SDK\Catalog\Category::getSeoTitle`
  - Added `\Wizaplace\SDK\Catalog\Category::getSeoDescription`
+ - Added `\Wizaplace\SDK\Seo\SeoService::listSlugs` and associated classes
 
 ### Bugfixes